feat: 16-field chained finalize with u32-packed result digest
The 8->16 field port was never finished on the aggregator finalize path:
the guest reader, input-gen validator and digest parser still assumed 8
fields while the host producer and circuit-primitives were at 16, so a
chained finalize hit the genesis-root assert.

Widen the finalize frame to 64-coord ballot / 16 results / 16 CP proofs.
The digest's publicValues is capped at 256 B (64 u32 words), so pack the
16 plaintext tallies as one u32 each into the existing 16-word results
slot (guest asserts each < 2^32) rather than 16 u64s. Rebuild the
aggregator ELF and refreeze CircuitRelease to the new agg/batch vks.

+++ go-sdk/chain/sequencer.go
@@ -289,7 +289,7 @@ func (s *Sequencer) Finalize(privKey *big.Int) (*FinalResult, error) {
 		return nil, err
 	}
 	for i, r := range results {
-		if digest.Results[i] != r {
+		if uint64(digest.Results[i]) != r {
 			return nil, fmt.Errorf("digest result[%d] = %d, want %d", i, digest.Results[i], r)
 		}
 	}
